Feat:Fine-tune GPT-3.5Turbo using YouTubeComments
Retreive the list of comments and responses from the YT_comments.csv and create message dictionary
Convert data to JSONL format for training and validation datasets since OpenAI model expects training_data to be in JSONL format.
Perform token counting using Token_Calculator.ipynb to make sure the number of tokens inside the training_data doesnt exceed the context length of the model.
Check and validate the format of training and validation data to ensure compliance with OpenAI fine-tuning requirements.
Upload the  both of the  validated datasets to OpenAI API storage.
Fine-tune the model with the prepared datasets.
Store the responses from the fine-tuned and non-fine-tuned models in Result.txt for comparison.